Run the server to check build before deploy

+++ clean_build.sh
@@ -1,9 +1,10 @@
 #!/bin/sh -e
 
-#LABEL=blog:$(git rev-parse --short origin/master)
-docker build --no-cache -t blog .
-docker run -ti -d blog /bin/bash
+docker build -t blog .
+docker run -dti -p 8080:8080 blog /bin/bash
+open http://localhost:8080/
 CONTAINER_ID=$(docker ps -alq)
-rm -r ./build
+if [ -d "./build" ]; then
+  rm -r ./build
+fi
 docker cp $CONTAINER_ID:/blog/build ./build
-docker stop $CONTAINER_ID
